refactor: Rename full-apc to expanded-apc
Fix a naming fib.

We have agreed on the following terminology: partial-apc refers to the
input to the partial-apc pipeline. Expanded-apc refers to the output of
the partial-apc pipeline. Full-apc refers to the message format that
include the full HFP details including the vehicle journey information.

Full-apc is sent directly from the vehicles just like partial-apc.
Expanded-apc should be identical to full-apc. That eases parsing and
data consumption in other parts of transitdata and HSL Data Warehouse.