fix: Workaround for eslint-plugin-import
There is a standstill in
https://github.com/import-js/eslint-plugin-import/issues/2751
due to the maintainer favouring backwards compatibility over new
features.

Unfortunately for us, that breaks @tsconfig/bases which relies on the
new array syntax for extends in tsconfig.json.

Using overrides is a workaround that is hopefully not needed later on.
It's painful to have to remember to remove the workaround but the
TypeScript train is moving on.